BF: move the Before Experiment JS code to precede window
Needed to allow @thewhodidthis to monkey-patch the PsychoJS lib within a
Code Component if needed before the window is created.

Downside is that now people using the Before Experiment block can't
reference the window or psychojs, but they can do that in Begin
Experiment and the Before Experiment is more for things before psychojs
exists anyway.