BF: Emotive Record causing standalone app not to load
When the app was bundled into ProgramFiles for a release (rather than in
the developers folder) the EmotivRecord Component was causing the app
not to load.

The issue was the logging, which we've fixed with belt and braces:
1. the logging file was set to store as 'cortex.log' but when the
Standalone app is launched from the Start menu that gets saved into
ProgramFiles and would require admin privs. FIX: set the file to log in
the psychopy AppData folder, which is always in user space
2. the reason this was an issue at launch time (when cortex should not
need to be logging) was because the Emotiv Record component was
importing the hardware code itself. This isn't needed by the app - only
load the hardware at run-time, not within the app