BF: recursion error trying to load Keyboard
I hav ewritten a check a long time ago to see if we needed permissions
to collect keyboard from Accessibility mode on Mac.

@sol recently moved that to be inside `Keyboard.__init__` but that
caused a recursion error f88fa3bd3e

BUT this shouldn't be needed now anyway, at least for the Standalone,
due to the code-signing and notarization process - we get the permission
from that. Question mark about what happens for dev users on mac