BF: Fixes #4135 - labels now work in DlgFromDict
In the most recent release version if you pass at labels argument to the
DlgFromDict class then when you return from the dialog it raises a KeyError.

This was caused by the field name in the dialog no longer matching the
dictionary key which is passed to the __init__ function. The fix was to store
the labels argument as an attribute on the dialog object to test against
when the dialog box returns.