BF: further fix to minolta str/bytes incompatibility
The lines :
```
        if message[-2:] != '\r\n':
            message += b'\r\n'  # append a newline if necess
```
The message would always get b'\r\n' added because it was being sent as
bytes so the comparison with '\r\n' would always be False and b'\r\n'
would then alwasy be added.
The confusion is partly because the messages were being sent in bytes
rather than in the natural string format for the python version (unicode).
It's easier to keep everything as the native string (now unicode) until
the last moment where we need btyes (`com.write()`) and then switch back
as soon as possible when reading. This way there are fewer places to
think about "is this a bytes object yet?"