RF: Remove iohub stroop_keyboard Builder demo
Removes the demos/builder/Feature Demos/iohub folder, including the stroop_keyboard project as it is no longer of use since Builder generated scripts use keyboard.Keyboard for kb event timing. The original purpose of the demo was to show how to access the iohub keyboard when Builder use to use event.getKeys() for event timing. This is no longer relevant.