ENH: Builder experiments now support updating values on QUEST
Staircases and QUEST classes support the case where the level
that was presented was not actually the level suggested by the
staircase by updating in the addResponse() method. This was not
being exposed in Builder boilerplate, however.

In this PR we add `level` back during addResponse so that if a
Code Component does something like:
```
if level > 1:
    level = 1
```
then the level value will be passed back correctly during addResponse
and stored by the QUEST/StairHandler.

The operation is in place for both Py/JS engines