FF: Adding placeholder _hw_handle to glfw backend
FF: Adding placeholder _hw_handle to glfw backend. Just setting to be a unique int value for each window, which is all that is needed for the new iohub multiwindow mouse functionality. Can be replaced with real _hw_handle number once we know how to get it in glfw.
RF: move ioHubConnection.unregisterWindowHandles() call from window backends to the Window class itself