BF: Modification of parameter with default
Fix series of the same LGTM.com error.

The default value of keyword arguments is mutable and is eventually
mutated, changing the default value in future function calls. Perhaps
it's currently not an issue in some cases, but it is a Python anti-
pattern that may cause hard to detect problems in the future.