Fix all other typos found by codespell
Note that this fixes both code comments and user-visible text, because
fixing only user-visible text and more importantly maintaining in the
long term is a difficult task - at least with codespell which doesn't
make a difference between code, code comments, documentation or
user-visible output.

If the above or the extent of the changes are a problem, please tell me,
I may be able to split into different commits.