PKG: rewrite of the Windows install script
fixes #3201 (and duplicate #3505): the uninstaller was always looking to
  uninstall from the default location rather than its actual location.
  This was due to NSIS confusing system of assigning variables during
  onInit. It turns out that doesn't need doing anyway - we can use the
  MultiUser settings MULTIUSER_USE_PROGRAMFILES64 and an precompiler !if
  to do this job.

User can now also choose NOT to uninstall and carry on (e.g. make 2 installs)

Permanently fixed the issue with long strings for environment variables
by dropping the use of EnvVarUpdate and using EnVar instead. (Previous fix
was to use a long-string version of NSIS which was easy to forget).