BF: preferences.generateSpec was no longer operating correctly as a script
It had been turned into a function for testing purposes by a408ec9ddf but
was no longer being used to generate the specs.

Also, the platforms had been manually changed to make PsychoPyDark their
default Theme. Dark is cool for the users that want to find it, but not
so good as a default, especially because doesn't quite apply to every
control (e.g. the status bar)