FF: for ResourceManager we don't want to "wait" resources just check
For StaticCompon we want to wait at the end (but the screen is static
anyway and this behaviour would be in keeping with PsychoPy behaviour,
whereas for ResourceManager we want to allow dynamic updating. In this
instance we should check but not freeze. User can see if this compon is
finished at any point