BF: if statements in Keyboard meant that event was the default backend
There were 3 if statements that all included `''` as one of the valid
options. So, if no backend was given, we ended up importing all 3
backend libs and then setting event as the one to use.

We want iohub to be default (I believe) so *either* use if...elif...elif
or set the other options not to include ''