FF: undoing previous commit (move elif back to be if) but move to setBackend
I now understnad the idea of using if so that if the PTB or iohub options
fail we still fallback to event. Just we needed event to be working! :-)

But there was another problem that using setEvent would fail because it
wasn't going through the steps used during init (e.g. to import iohub)