BF: Builder JS boilerplate now honors the loop.isTrials setting
We previously used a single function endLoopIteration that was general
to all loops. This commit changes that so that each loop has its own
endIteration function and then we can decide which of those endIteration
functions call nextEntry. This is easier than trying to work out whether
this loop isTrials at run-time