DOCS: Fixes to styling (of the bullet lists)
Several files were being defined in docs/source/_static (which is really
just for site-specific changes) rather than in themes/psychopy/static
which is for other subdomains using the same theme. css files should be
in themes. Sphinx gives priority to the site-specific files and so some
fixes we had made to the theme were not being applied here, like the fix
to the blockquote left-hand border

Second indent of a list was showing up with a block-quote left hand border
which looked really odd. This had been fixed in one location
  themes/static/psychopy.css
but was broken in which was taking precendent
  source/_static/psychopy.css

This commit removed source/_static/psychopy.css to avoid confusion and
applies some of the other fixes back to themes/static instead