TextBox2 Color Support
An enhancement and set of bugfixes for TextBox2.

The main enhancement is giving users to add custom colors using a new tag, <c>. The format is: <c=#ffffff>X</c>, where the color information is given in 6-digit hex format. 'c' was used rather than 'color' to keep raw text as small as possible, however, an alias between 'c' and 'color' can be easily added. The current implementation allows each individual character to be colored independently. The default font color will be used when a color is not explicitly defined.

A further enhancement is the ability for users to apply multiple styles simultaneously (currently <b>, <i>, <c=>). The previous code had some logic that seemed to promote simultaneous <b><i>, however, it was not fully functional. Also, there were various bugs in the previous style implementation that have been fixed/changed. For example, previously, <i><i>X</i></i> was equivalent to <b>X</b> since the value ITALIC = 1 and BOLD = 2 and ITALIC + ITALIC = BOLD. These issues are no longer present since the _style list has been replaced with a new class Style(). This is a collection of lists that store individual style information per-character. The list includes necessary functions in order to require minimal modifications to the editing features.

Text is still editable and behaves exactly as it did before. All styles are inherited from the character to the left of the caret (even though I think in some cases it should inherit from the right as well).

Using a modified 'compare_text_timing.py', I found that the changing text is slightly slower <1ms. However, I think there's a lot of room for improvement by using numpy arrays. Also there are currently regex calls for color which may be eventually replaced with a more efficient parser.