BF: if framerate is variable don't keep trying to measure
The code for moviepy and opencv backends was checking for a framerate
and if didn't find one asked Window to make a measurement. BUT if the
issue was that the system timing was a bit shaky, then this meant that
every time you load a new movie it does another (failing) timing check.

Basically it makes matters much worse to keep trying to recheck timing.

Just assume it's 60Hz and be done with it!