PKG: actually, better to move building to private devops repo
Although github has ways to prevent secrets (keys) being accidentally revealed
that can be circumvented by a malicious actor.  So it's better to move
to a private repo that only trusted individuals have access to.

The workarounds to reveal keys have not been used, so
the keys are all currently safe and have been
removed so that they cannot be accessed in future from here