Minor textbox fixes
Previously when size and textgrid_shape were both None, there would be a crash. I fixed this by detecting when both were none and setting textgrid_shape to len(text) by 1, thus solving the issue. I also fixed a problem where size was too small, which caused textgrid.py to crash. I added a check to make sure the size wouldn't lead to an integer division error. Instead, if the size is not large enough (e.g. would cause an error), a ValueError is raised which tells the user to increase the size or use the textgrid_shape option.