BF: sound.pause should not be a toggle
Toggling was causing sounds to restart in at the end of Routines where
they had ended or had been stopped.

It was set to toggle, only for this sound backend, in 2023.1 but that was
only documented in 2023.2.0 and I think we should consider it a bug
rather than a feature.