BF: for institutions with a proxy server set this before connecting to pav
Not many institutions these days have proxy servers but those that do
will block access to gitlab as a result. We need to set up the proxy (as
set in the user prefs) and then hand that requests.Session to the
gitlab lib.

Previously we were creating the requests session after the gitlab
connection and not using th proxy for this at all