BF show start parameter for stop only eyetracker components
Currently the Stop Only action type causes experiments to crash. This is because the stop recording component never has its status set to START in the compiled python code. It does make sense to have a Start time because in theory you might want to Stop following a stimulus or something like that. This edit means that the stop recording component is set to start and therefore the recording is stopped.