BF: Prevent _legacy on js compiled file
When a psychopy builder file has been synced to pavlovia, the pavlovia project creates the url run.pavlovia.org/username/projectname where projectname is the name of the js file it is looking for. If you use Use Version then at the moment it compiles projectname_legacy.js - which means that file is not used by the pavlovia project. We need the js file that is compiled to always be the same name as the .psyexp file we are working on and syncing from, even if a different version is used.