BF: Types mismatch caused SetThreadPriority to fail on win32
I noticed rush consistently failing during testing,
and it seemed that SetThreadPriority was getting an invalid
handle. The priority class would change, but the extra +2
bump would be missing (checked via the Process Explorer tool)

This adds types to all args/return values for win32 functions, which
seems to help SetThreadPriority succeed. However, I haven't been able
to coax the "realtime" settings out-- even though it visits the right
branch, it still lands on the high-priority-but-not-realtime settings...