BF: Enable pyglet window on separate X screen on linux.
If a machine has been configured to have separate X screens
(e.g. for use with psychtoolbox), the correct screen can only be selected
by pyglet if the screen number is passed to the Display constructor.

However, the default case for Ubuntu is that all the physical displays
are mapped to the same X screen, and so requesting a higher screen number at that step
would fail early (see e.g. https://github.com/psychopy/psychopy/issues/5888).

This hopefully deals with both cases, but the separate-X-screen
case may only work if pyglet's shadow window is also disabled?
gluNewTess init has been moved later in the process to allow the shadow
window to be disabled.