Delay the populate actions of dependent parameters
This commit is intended to address issue #7173. In sum, the bug happens
when there are two parameters that have inputType = "choice" and have also
declared dependsOn relationships with a third parameter.

In the current state of the code, only one of the two dependent parameters
has its list repopulated. This happens because, in the loop over the
declared dependencies (self.parent.element.depends) for a given element (in
method CategoryPage.checkDepends), the value of dependencyParam.val is
prematurely set.

With this commit, the repopulation, as well as the setting of
dependencyParam.val, are delayed after the for loop completes.

I hope that this solutions will be acceptable for the upstream authors.